Refactor RepositoryContext and RepositoryProvider to streamline file status management and enhance error handling. Introduced new hooks for repository operations and file status, improving code organization and reducing redundancy. Updated imports and removed unused code for better clarity.
