+++ apps/figma-plugin/src/context/Repository/RepositoryProvider.tsx
@@ -1,23 +1,29 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { RepositoryContext } from './RepositoryContext';
-import { useFigma, useFileData, useRepositoryInstance, useAdapterWorker } from '../../hooks';
+import {
+  useFigma,
+  useFileData,
+  useRepositoryInstance,
+  useAdapterWorker,
+  useRepositoryError,
+  useFileStatus,
+  useRemoteFiles,
+  useRepositoryOperations,
+  useProjectValidation,
+  FileStatus,
+} from '../../hooks';
 import type {
-  RecursicaConfiguration,
   RecursicaVariablesSchema,
   CollectionToken,
   VariableReferenceValue,
   Token,
 } from '@recursica/schemas';
-import { type UserInfo, type Project, type CommitAction } from '../../services/repository';
+import { type UserInfo, type Project } from '../../services/repository';
 import { isColorOrFloatToken } from '@recursica/common';
 
 interface AdapterFile {
   path: string;
   content: string;
-}
-
-function cleanPath(path: string) {
-  return path.replace(/^\/+/, '');
 }
 
 // Helper function to normalize variable names
@@ -138,57 +144,11 @@ function generateCSSFromBundledJson(bundledJson: string): string {
   }
 }
 
-export enum FileStatus {
-  Error,
-  Pending,
-  Loading,
-  Done,
-}
-
-export interface Status {
-  quantity: number;
-  status: FileStatus;
-}
-
-export interface FilesStatus {
-  icons: Status;
-  uiKit: Status;
-  tokens: Status;
-  themes: Status;
-  adapter: Status;
-}
-
-const INITIAL_FILES_STATUS: FilesStatus = {
-  icons: { quantity: 0, status: FileStatus.Pending },
-  uiKit: { quantity: 0, status: FileStatus.Pending },
-  tokens: { quantity: 0, status: FileStatus.Pending },
-  themes: { quantity: 0, status: FileStatus.Pending },
-  adapter: { quantity: 0, status: FileStatus.Pending },
-};
-
-export enum ValidationStatus {
-  Valid,
-  Invalid,
-  NotSelected,
-}
-
-export interface RepositoryError {
-  message: string;
-  details?: string;
-  code?: string;
-  timestamp: number;
-}
-
 export function RepositoryProvider({ children }: { children: React.ReactNode }) {
   const [userInfo, setUserInfo] = useState<UserInfo | null>(null);
   const [selectedProjectId, setselectedProjectId] = useState<string | undefined>(undefined);
   const [userProjects, setUserProjects] = useState<Project[]>([]);
   const [prLink, setPrLink] = useState<string | null>(null);
-  const [filesStatus, setFilesStatus] = useState<FilesStatus>(INITIAL_FILES_STATUS);
-  const [validationStatus, setValidationStatus] = useState<ValidationStatus>(
-    ValidationStatus.NotSelected
-  );
-  const [error, setError] = useState<RepositoryError | null>(null);
   const initConfig = useRef<boolean>(false);
 
   const selectedProject = useMemo(() => {
@@ -205,134 +165,35 @@ export function RepositoryProvider({ children }: { children: React.ReactNode }) 
     useFileData(selectedProject);
   const { repositoryInstance } = useRepositoryInstance();
   const { runAdapter } = useAdapterWorker();
+  const { error, setRepositoryError, clearError, isExpectedError, getErrorMessage } =
+    useRepositoryError();
+  const {
+    filesStatus,
+    updateFilesStatus,
+    updateAdapterStatus,
+    updateAllStatusesToDone,
+    resetFilesStatus,
+  } = useFileStatus();
+  const { validationStatus, validateProject, resetValidationStatus } = useProjectValidation();
+  const {
+    fetchUserInfo: fetchUserInfoOp,
+    getUserProjects: getUserProjectsOp,
+    createBranch,
+    getConfig,
+    commitFiles,
+    createPullRequest,
+  } = useRepositoryOperations();
 
-  // Fetch remote files when repository is available
-  useEffect(() => {
-    const fetchRemoteFiles = async () => {
-      if (!repositoryInstance || !selectedProject) return;
+  // Use remote files hook
+  useRemoteFiles(repositoryInstance, selectedProject, setRemoteVariablesJson, setRemoteIconsJson);
 
-      try {
-        // Fetch remote icons
-        const iconsFilename = 'recursica-icons.json';
-        try {
-          const iconsFileRaw = await repositoryInstance.getSingleFile(
-            selectedProject,
-            iconsFilename,
-            selectedProject.defaultBranch
-          );
-          const remoteIconsJson = JSON.stringify(iconsFileRaw);
-          setRemoteIconsJson(remoteIconsJson);
-        } catch (error) {
-          if (error instanceof Error && error.message.includes('404')) {
-            setRemoteIconsJson('{}');
-          } else {
-            console.error('Error fetching remote icons:', error);
-          }
-        }
-
-        // Fetch remote variables
-        const bundledFilename = 'recursica-bundle.json';
-        try {
-          const bundledFileRaw = await repositoryInstance.getSingleFile<RecursicaVariablesSchema>(
-            selectedProject,
-            bundledFilename,
-            selectedProject.defaultBranch
-          );
-          const remoteBundledJson = JSON.stringify(bundledFileRaw);
-          setRemoteVariablesJson(remoteBundledJson);
-        } catch (error) {
-          if (error instanceof Error && error.message.includes('404')) {
-            setRemoteVariablesJson('{}');
-          } else {
-            console.error('Error fetching remote variables:', error);
-          }
-        }
-      } catch (error) {
-        console.error('Error fetching remote files:', error);
-      }
-    };
-
-    fetchRemoteFiles();
-  }, [repositoryInstance, selectedProject]);
-
-  // Error handling helper functions
-  const setRepositoryError = (message: string, details?: string, code?: string) => {
-    const error: RepositoryError = {
-      message,
-      details,
-      code,
-      timestamp: Date.now(),
-    };
-    setError(error);
-    console.error('Repository Error:', error);
-  };
-
-  const clearError = () => {
-    setError(null);
-  };
-
-  const isExpectedError = (error: unknown): boolean => {
-    if (error instanceof Error) {
-      // 404 errors are expected for missing files
-      if (error.message.includes('404')) return true;
-      // Network errors that might be temporary
-      if (error.message.includes('Network Error') || error.message.includes('fetch')) return true;
-    }
-
-    // Check for axios errors
-    if (error && typeof error === 'object' && 'response' in error) {
-      const axiosError = error as { response: { status: number } };
-      const status = axiosError.response.status;
-      // 404 errors are expected for missing files
-      if (status === 404) return true;
-      // 409 conflicts are expected for existing PRs
-      if (status === 409) return true;
-      // 422 validation errors might be expected in some cases
-      if (status === 422) return true;
-    }
-
-    return false;
-  };
-
-  const getErrorMessage = (error: unknown): string => {
-    if (error && typeof error === 'object' && 'response' in error) {
-      const axiosError = error as {
-        response: {
-          status: number;
-          data?: { message?: string };
-        };
-      };
-      const status = axiosError.response.status;
-      switch (status) {
-        case 401:
-          return 'Authentication failed. Please check your access token.';
-        case 403:
-          return 'Access denied. You may not have permission to access this repository.';
-        case 404:
-          return 'Resource not found. The repository or file may not exist.';
-        case 429:
-          return 'Rate limit exceeded. Please try again later.';
-        case 500:
-          return 'Server error. Please try again later.';
-        case 503:
-          return 'Service temporarily unavailable. Please try again later.';
-        default:
-          return axiosError.response.data?.message || `HTTP ${status} error occurred.`;
-      }
-    }
-
-    if (error instanceof Error) {
-      if (error.message.includes('Network Error')) {
-        return 'Network connection error. Please check your internet connection.';
-      }
-      if (error.message.includes('timeout')) {
-        return 'Request timed out. Please try again.';
-      }
-      return error.message;
-    }
-
-    return 'An unexpected error occurred.';
-  };
+  const updateSelectedProjectId = useCallback(
+    (projectId: string) => {
+      setselectedProjectId(projectId);
+      updateSelectedProject(projectId);
+    },
+    [updateSelectedProject]
+  );
 
   useEffect(() => {
     if (repository) {
@@ -341,82 +202,68 @@ export function RepositoryProvider({ children }: { children: React.ReactNode }) 
   }, [repository]);
 
   useEffect(() => {
+    const fetchUserInfo = async () => {
+      if (!repositoryInstance) return;
+      try {
+        const data = await fetchUserInfoOp(repositoryInstance);
+        if (data) {
+          setUserInfo(data);
+          clearError();
+        }
+      } catch (error) {
+        if (!isExpectedError(error)) {
+          setRepositoryError(
+            'Failed to fetch user information',
+            getErrorMessage(error),
+            'USER_INFO_FETCH_ERROR'
+          );
+        }
+      }
+    };
+
     if (repositoryInstance) {
       fetchUserInfo();
     }
-  }, [repositoryInstance]);
+  }, [repositoryInstance, fetchUserInfoOp]);
 
   useEffect(() => {
+    const getUserProjects = async () => {
+      if (!repositoryInstance) return;
+      try {
+        const projects = await getUserProjectsOp(repositoryInstance);
+        setUserProjects(projects);
+        clearError();
+      } catch (error) {
+        if (!isExpectedError(error)) {
+          setRepositoryError(
+            'Failed to fetch user projects',
+            getErrorMessage(error),
+            'PROJECTS_FETCH_ERROR'
+          );
+        }
+      }
+    };
+
     if (userInfo && repositoryInstance) {
       getUserProjects();
     }
-  }, [userInfo, repositoryInstance]);
+  }, [userInfo, repositoryInstance, getUserProjectsOp]);
 
   // Auto-select project if there's only one available and no project is currently selected
   useEffect(() => {
     if (userProjects.length === 1 && !selectedProjectId) {
       updateSelectedProjectId(userProjects[0].id);
     }
-  }, [userProjects, selectedProjectId]);
-
-  const fetchUserInfo = async () => {
-    if (!repositoryInstance) return;
-
-    try {
-      const data = await repositoryInstance.getUserInfo();
-      setUserInfo(data);
-      clearError(); // Clear any previous errors on success
-    } catch (error) {
-      if (!isExpectedError(error)) {
-        setRepositoryError(
-          'Failed to fetch user information',
-          getErrorMessage(error),
-          'USER_INFO_FETCH_ERROR'
-        );
-      }
-    }
-  };
-
-  const getUserProjects = async () => {
-    if (!repositoryInstance) return;
-    try {
-      const projects = await repositoryInstance.getUserProjects();
-      setUserProjects(projects);
-      clearError(); // Clear any previous errors on success
-    } catch (error) {
-      if (!isExpectedError(error)) {
-        setRepositoryError(
-          'Failed to fetch user projects',
-          getErrorMessage(error),
-          'PROJECTS_FETCH_ERROR'
-        );
-      }
-    }
-  };
+  }, [userProjects, selectedProjectId, updateSelectedProjectId]);
 
   useEffect(() => {
-    if (selectedProject) {
-      validateProject();
-    }
-  }, [selectedProject]);
-
-  const validateProject = async (): Promise<boolean> => {
-    if (!repositoryInstance) return false;
-    try {
-      const config = await getConfig(selectedProject?.defaultBranch || '');
-      setValidationStatus(
-        config.project !== undefined ? ValidationStatus.Valid : ValidationStatus.Invalid
-      );
-      return config.project !== undefined;
-    } catch (error) {
-      if (error instanceof Error && error.message.includes('404')) {
-        console.error('Project is not valid');
-        setValidationStatus(ValidationStatus.Invalid);
-        return false;
+    const validateProjectAsync = async () => {
+      if (selectedProject) {
+        await validateProject(repositoryInstance, selectedProject, getConfig);
       }
-      throw error;
-    }
-  };
+    };
+    validateProjectAsync();
+  }, [selectedProject, validateProject, repositoryInstance, getConfig]);
 
   const initializeRepo = async () => {
     if (!repositoryInstance) return;
@@ -472,34 +319,6 @@ export function RepositoryProvider({ children }: { children: React.ReactNode }) 
     }
   };
 
-  const updateFilesStatus = (vars: string | null, icons: string | null) => {
-    if (vars) {
-      const bundledFileRaw = JSON.parse(vars);
-      setFilesStatus((prev) => ({
-        ...prev,
-        tokens: {
-          quantity: Object.keys(bundledFileRaw.tokens).length,
-          status: FileStatus.Loading,
-        },
-        themes: {
-          quantity: Object.keys(bundledFileRaw.themes).length,
-          status: FileStatus.Loading,
-        },
-        uiKit: {
-          quantity: Object.keys(bundledFileRaw.uiKit).length,
-          status: FileStatus.Loading,
-        },
-      }));
-    }
-    if (icons) {
-      const iconsFileRaw = JSON.parse(icons);
-      setFilesStatus((prev) => ({
-        ...prev,
-        icons: { quantity: Object.keys(iconsFileRaw).length, status: FileStatus.Loading },
-      }));
-    }
-  };
-
   useEffect(() => {
     if (fileLoadingData.bundledJson || fileLoadingData.iconsJson) {
       updateFilesStatus(fileLoadingData.bundledJson, fileLoadingData.iconsJson);
@@ -513,8 +332,17 @@ export function RepositoryProvider({ children }: { children: React.ReactNode }) 
     }
 
     try {
-      const targetBranch = await createBranch(selectedProject);
-      const config = await getConfig(targetBranch);
+      const targetBranch = await createBranch(
+        repositoryInstance,
+        selectedProject,
+        userInfo?.username
+      );
+      const config = await getConfig(
+        repositoryInstance,
+        selectedProject,
+        targetBranch,
+        initConfig.current
+      );
 
       // Run adapter with proper error handling
       let adapterFiles: AdapterFile[] = [];
@@ -530,20 +358,11 @@ export function RepositoryProvider({ children }: { children: React.ReactNode }) 
           fileLoadingData,
           (status) => {
             if (status === 'loading') {
-              setFilesStatus((prev) => ({
-                ...prev,
-                adapter: { quantity: 1, status: FileStatus.Loading },
-              }));
+              updateAdapterStatus(FileStatus.Loading, 1);
             } else if (status === 'done') {
-              setFilesStatus((prev) => ({
-                ...prev,
-                adapter: { ...prev.adapter, status: FileStatus.Done },
-              }));
+              updateAdapterStatus(FileStatus.Done);
             } else if (status === 'error') {
-              setFilesStatus((prev) => ({
-                ...prev,
-                adapter: { ...prev.adapter, status: FileStatus.Error },
-              }));
+              updateAdapterStatus(FileStatus.Error, 0);
             }
           }
         );
@@ -555,10 +374,7 @@ export function RepositoryProvider({ children }: { children: React.ReactNode }) 
           'ADAPTER_EXECUTION_ERROR'
         );
         // Update files status to show error immediately
-        setFilesStatus((prev) => ({
-          ...prev,
-          adapter: { ...prev.adapter, status: FileStatus.Error },
-        }));
+        updateAdapterStatus(FileStatus.Error);
         throw error; // Re-throw to prevent further execution
       }
 
@@ -573,7 +389,31 @@ export function RepositoryProvider({ children }: { children: React.ReactNode }) 
         }
       }
 
-      await commitFiles(targetBranch, adapterFiles, config);
+      await commitFiles(
+        repositoryInstance,
+        selectedProject,
+        targetBranch,
+        adapterFiles,
+        config,
+        {
+          bundledJson: fileLoadingData.bundledJson,
+          iconsJson: fileLoadingData.iconsJson,
+          bundledFilename: fileLoadingData.bundledFilename,
+          iconsFilename: fileLoadingData.iconsFilename,
+        },
+        initConfig.current
+      );
+
+      updateAllStatusesToDone();
+
+      const pullRequestUrl = await createPullRequest(
+        repositoryInstance,
+        selectedProject,
+        targetBranch
+      );
+      if (pullRequestUrl) {
+        setPrLink(pullRequestUrl);
+      }
     } catch (error) {
       // Ensure any error is properly handled and UI is updated
       if (!isExpectedError(error)) {
@@ -587,167 +427,15 @@ export function RepositoryProvider({ children }: { children: React.ReactNode }) 
     }
   };
 
-  const getConfig = async (targetBranch: string): Promise<RecursicaConfiguration> => {
-    if (!repositoryInstance) throw new Error('Failed to get repository instance');
-    if (!selectedProject) throw new Error('Failed to get selected project');
-    if (initConfig.current) {
-      return {
-        project: {
-          name: selectedProject.name,
-          root: '',
-        },
-      };
-    }
-    const configFile = await repositoryInstance.getSingleFile<RecursicaConfiguration>(
-      selectedProject,
-      'recursica.json',
-      targetBranch
-    );
-    if (!configFile) throw new Error('Failed to get config file');
-    return configFile;
-  };
-
-  const commitFiles = async (
-    targetBranch: string,
-    adapterFiles: AdapterFile[],
-    config: RecursicaConfiguration
-  ): Promise<void> => {
-    if (!repositoryInstance || !selectedProjectId || !userInfo) {
-      throw new Error('Repository instance, project ID, or user info not available');
-    }
-
-    if (!selectedProject) {
-      throw new Error('Selected project is not available');
-    }
-
-    let rootPath = '';
-    if (typeof config.project === 'object' && config.project.root) {
-      rootPath = config.project.root;
-    }
-    let variablesFilename = fileLoadingData.bundledFilename;
-    if (rootPath) {
-      variablesFilename = rootPath + '/' + variablesFilename;
-    }
-    let svgIconsFilename = fileLoadingData.iconsFilename;
-    if (rootPath) {
-      svgIconsFilename = rootPath + '/' + svgIconsFilename;
-    }
-
-    const actions: CommitAction[] = [];
-    if (fileLoadingData.bundledJson) {
-      const exists = await repositoryInstance.fileExists(
-        selectedProject,
-        variablesFilename,
-        targetBranch
-      );
-      actions.push({
-        action: exists ? 'update' : 'create',
-        file_path: variablesFilename,
-        content: fileLoadingData.bundledJson,
-      });
-    }
-
-    if (initConfig.current) {
-      actions.push({
-        action: 'create',
-        file_path: 'recursica.json',
-        content: JSON.stringify(config, null, 2),
-      });
-    }
-
-    if (fileLoadingData.iconsJson) {
-      const exists = await repositoryInstance.fileExists(
-        selectedProject,
-        svgIconsFilename,
-        targetBranch
-      );
-      actions.push({
-        action: exists ? 'update' : 'create',
-        file_path: svgIconsFilename,
-        content: fileLoadingData.iconsJson,
-      });
-    }
-
-    if (adapterFiles.length > 0) {
-      for (const file of adapterFiles) {
-        if (!selectedProject) {
-          throw new Error('Selected project is not available during file processing');
-        }
-        const exists = await repositoryInstance.fileExists(
-          selectedProject,
-          cleanPath(file.path),
-          targetBranch
-        );
-        actions.push({
-          action: exists ? 'update' : 'create',
-          file_path: cleanPath(file.path),
-          content: file.content,
-        });
-      }
-    }
-
-    await repositoryInstance.commitFiles(
-      selectedProject,
-      targetBranch,
-      `Files committed by Recursica\n${variablesFilename}`,
-      actions
-    );
-
-    setFilesStatus((prev) => ({
-      ...prev,
-      icons: { ...prev.icons, status: FileStatus.Done },
-      uiKit: { ...prev.uiKit, status: FileStatus.Done },
-      tokens: { ...prev.tokens, status: FileStatus.Done },
-      themes: { ...prev.themes, status: FileStatus.Done },
-    }));
-
-    try {
-      const pullRequest = await repositoryInstance.createPullRequest(
-        selectedProject,
-        targetBranch,
-        selectedProject.defaultBranch,
-        'New recursica tokens release'
-      );
-      setPrLink(pullRequest.url);
-    } catch (error) {
-      console.error('Failed to create pull request, trying to find existing one:', error);
-
-      // Check if there's already an open PR/MR for this branch
-      const hasOpenPR = await repositoryInstance.hasOpenPullRequest(
-        selectedProject,
-        targetBranch,
-        selectedProject.defaultBranch
-      );
-
-      if (hasOpenPR) {
-        console.log('Pull request already exists for this branch');
-      }
-    }
-  };
-
-  const createBranch = async (project: Project) => {
-    if (!repositoryInstance) throw new Error('Failed to create branch');
-    const branch = await repositoryInstance.createBranch(
-      project,
-      `recursica-${userInfo?.username}-${Date.now()}`,
-      project.defaultBranch
-    );
-    return branch.name;
-  };
-
-  const updateSelectedProjectId = (projectId: string) => {
-    setselectedProjectId(projectId);
-    updateSelectedProject(projectId);
-  };
-
   const resetRepository = async () => {
     setPrLink(null);
-    setError(null); // Clear any existing errors
+    clearError(); // Clear any existing errors
     initConfig.current = false;
-    setFilesStatus(INITIAL_FILES_STATUS);
+    resetFilesStatus();
+    resetValidationStatus();
     // Clear remote file data
     clearRemoteData();
-    return await validateProject();
+    return await validateProject(repositoryInstance, selectedProject, getConfig);
   };
 
   const value = {
