+++ apps/figma-plugin/src/hooks/index.ts
@@ -3,3 +3,8 @@ export { useRepository } from './useRepository';
 export { useFileData } from './useFileData';
 export { useRepositoryInstance } from './useRepositoryInstance';
 export { useAdapterWorker } from './useAdapterWorker';
+export { useRepositoryError, type RepositoryError } from './useRepositoryError';
+export { useFileStatus, type FilesStatus, FileStatus } from './useFileStatus';
+export { useRemoteFiles } from './useRemoteFiles';
+export { useRepositoryOperations } from './useRepositoryOperations';
+export { useProjectValidation, ValidationStatus } from './useProjectValidation';
